added/fixed field length counters
Conflicts:
	miso-web/src/main/java/uk/ac/bbsrc/tgac/miso/webapp/controller/EditPlateController.java
	miso-web/src/main/webapp/pages/editBox.jsp
