+++ miso-web/src/main/webapp/pages/editIlluminaPool.jsp
@@ -105,7 +105,7 @@
   </tr>
   <tr>
     <td class="h">Pool Alias:</td>
-    <td><form:input path="alias"/></td>
+    <td><form:input id="alias" path="alias"/><span id="aliasCounter" class="counter"></span></td>
   </tr>
   <tr>
     <td class="h">Desired Concentration</td>
@@ -346,6 +346,17 @@
   Utils.timer.typewatchFunc(jQuery('#searchDilution'), function () {
     Pool.search.poolSearchLibraryDilution(jQuery('#searchDilution'), 'ILLUMINA')
   }, 300, 2);
+</script>
+
+<script type="text/javascript">
+  jQuery(document).ready(function () {
+    jQuery('#alias').simplyCountable({
+      counter: '#aliasCounter',
+      countType: 'characters',
+      maxCount: ${maxLengths['alias']},
+      countDirection: 'down'
+    });
+  });
 </script>
 
 <%@ include file="adminsub.jsp" %>
