+++ miso-web/src/main/webapp/pages/editLibrary.jsp
@@ -223,14 +223,14 @@
   <td>
     <c:choose>
       <c:when test="${not empty autogeneratedLibraryAlias}">
-        <form:input path="alias" value="${autogeneratedLibraryAlias}" class="validateable"/>
+        <form:input id="alias" path="alias" value="${autogeneratedLibraryAlias}" class="validateable"/>
       </c:when>
       <c:otherwise>
-        <form:input path="alias" class="validateable"/>
+        <form:input id="alias" path="alias" class="validateable"/>
       </c:otherwise>
     </c:choose>
 
-    <span id="aliascounter" class="counter"></span>
+    <span id="aliasCounter" class="counter"></span>
   </td>
 </tr>
 <tr>
@@ -246,8 +246,8 @@
       </c:otherwise>
     </c:choose>
     --%>
-    <form:input path="description" class="validateable"/>
-    <span id="descriptioncounter" class="counter"></span></td>
+    <form:input id="description" path="description" class="validateable"/>
+    <span id="descriptionCounter" class="counter"></span></td>
 </tr>
 <tr>
   <td class="h">Creation date:</td>
@@ -837,15 +837,6 @@
             "aaSorting": [
               [1, 'asc'],
               [3, 'asc']
-            ],
-            "aoColumns": [
-              null,
-              { "sType": 'natural' },
-              null,
-              null,
-              null,
-              null
-              <sec:authorize access="hasRole('ROLE_ADMIN')">, null</sec:authorize>
             ],
             "iDisplayLength": 50,
             "bJQueryUI": true,
@@ -1464,6 +1455,24 @@ function submitBulkLibraries() {
 </c:if>
 </div>
 </div>
+
+<script type="text/javascript">
+  jQuery(document).ready(function () {
+    jQuery('#alias').simplyCountable({
+      counter: '#aliasCounter',
+      countType: 'characters',
+      maxCount: ${maxLengths['alias']},
+      countDirection: 'down'
+    });
+
+    jQuery('#description').simplyCountable({
+      counter: '#descriptionCounter',
+      countType: 'characters',
+      maxCount: ${maxLengths['description']},
+      countDirection: 'down'
+    });
+  });
+</script>
 
 <%@ include file="adminsub.jsp" %>
