+++ miso-web/src/main/webapp/pages/editPool.jsp
@@ -124,7 +124,7 @@
   </tr>
   <tr>
     <td class="h">Alias:*</td>
-    <td><form:input id="alias" path="alias"/></td>
+    <td><form:input id="alias" path="alias"/><span id="aliasCounter" class="counter"></span></td>
   </tr>
   <tr>
     <td>Platform Type:</td>
@@ -450,6 +450,17 @@
   <%--}, 300, 2);--%>
   <%--</c:otherwise>--%>
   <%--</c:choose>--%>
+</script>
+
+<script type="text/javascript">
+  jQuery(document).ready(function () {
+    jQuery('#alias').simplyCountable({
+      counter: '#aliasCounter',
+      countType: 'characters',
+      maxCount: ${maxLengths['alias']},
+      countDirection: 'down'
+    });
+  });
 </script>
 
 <%@ include file="adminsub.jsp" %>
