+++ miso-web/src/main/webapp/pages/editSolidPool.jsp
@@ -103,7 +103,7 @@
                 </tr>
                 <tr>
                     <td class="h">Pool Alias:</td>
-                    <td><form:input path="alias"/></td>
+                    <td><form:input id="alias" path="alias"/><span id="aliasCounter" class="counter"></span></td>
                 </tr>
                 <tr>
                     <td class="h">Concentration</td>
@@ -315,6 +315,17 @@
 <script type="text/javascript">
     Utils.ui.addMaxDatePicker("creationDate", 0);
     Utils.timer.typewatchFunc(jQuery('#searchDilution'), function(){Pool.search.poolSearchEmPcrDilution(jQuery('#searchDilution'), 'SOLID')}, 300, 2);
+</script>
+
+<script type="text/javascript">
+  jQuery(document).ready(function () {
+    jQuery('#alias').simplyCountable({
+      counter: '#aliasCounter',
+      countType: 'characters',
+      maxCount: ${maxLengths['alias']},
+      countDirection: 'down'
+    });
+  });
 </script>
 
 <%@ include file="adminsub.jsp" %>
